auth-service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
passlib[bcrypt]
bcrypt==4.0.1
python-dotenv
pyjwt
python-jose
python-multipart